+++ trunk/wxMusik/src/Frames/Prefs/OptionFileAssocPanel.cpp
@@ -61,7 +61,7 @@ void OptionFileAssocPanel::DoFileAssociations()
 {
 
     bool bNotifyExplorer = false;
-    for(size_t i = 0; i < chklbFileAssocs->GetCount() ; i++)
+    for(unsigned int i = 0; i < chklbFileAssocs->GetCount() ; i++)
     {
         wxString sEntry =	chklbFileAssocs->GetString(i);
         wxArrayString arrElements;
